Login with overlay implemented.
