+++ MvpApi.Uwp/Dialogs/LoginDialog.xaml
@@ -5,8 +5,8 @@
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
-    SecondaryButtonText="cancel"
-    SecondaryButtonClick="LoginDialog_OnSecondaryButtonClick">
+    SecondaryButtonText="cancel">
+    <!--SecondaryButtonClick="LoginDialog_OnSecondaryButtonClick"
 
     <ContentDialog.PrimaryButtonStyle>
         <Style TargetType="Button">
@@ -20,5 +20,5 @@
 
         <WebView x:Name="LoginWebView"
                  LoadCompleted="WebView_OnLoadCompleted" />
-    </Grid>
+    </Grid>-->
 </ContentDialog>
