+++ MvpApi.Uwp/ViewModels/ShellPageViewModel.cs
@@ -15,7 +15,6 @@ using MvpApi.Common.Extensions;
 using MvpApi.Common.Models;
 using MvpApi.Services.Apis;
 using MvpApi.Services.Utilities;
-using MvpApi.Uwp.Dialogs;
 using MvpApi.Uwp.Helpers;
 using Newtonsoft.Json;
 using Template10.Common;
@@ -45,33 +44,25 @@ namespace MvpApi.Uwp.ViewModels
                 return;
             }
         }
-
-        /// <summary>
-        /// File path to locally saved MVP profile image
-        /// </summary>
+        
         public string ProfileImagePath
         {
             get => _profileImagePath;
             set
             {
-                //enforcing propChanged
                 _profileImagePath = value;
+
+                // The file path may be the same, but we still want the image to be reloaded whenever this is set
                 RaisePropertyChanged();
             }
         }
-
-        /// <summary>
-        /// Currently signed in MVP profile
-        /// </summary>
+        
         public ProfileViewModel Mvp
         {
             get => _mvp;
             set => Set(ref _mvp, value);
         }
         
-        /// <summary>
-        /// Denotes whether the user is currently logged in and able to make successful requests to the API
-        /// </summary>
         public bool IsLoggedIn
         {
             get => _isLoggedIn;
@@ -84,179 +75,18 @@ namespace MvpApi.Uwp.ViewModels
             set => Set(ref _webViewUri, value);
         }
 
-        /// <summary>
-        /// Logs in the user by showing a ContentDialog for authentication.
-        /// Silent Operation if the user has previously logged in, a silent login will be attempted first (using the refresh token).
-        /// </summary>
-        /// <returns></returns>
-        public async Task<bool> SignInAsync()
-        {
-            if (IsLoggedIn)
-            {
-                return IsLoggedIn;
-            }
-            else
-            {
-                await SignOutAsync();
-                return IsLoggedIn;
-            }
-            // Initialize LoginDialog. This is a special ContentDialog that will handle authentication lifecycle
-            //if(_loginManager == null)
-            //{
-            //    _loginManager = new LoginDialog();
-            //}
 
-            //try
-            //{
-            //    // Invoke sign in procedure (this handles token storage internally)
-            //    await _loginManager.AttemptSilentRefreshAsync();
-
-            //    if (!string.IsNullOrEmpty(_loginManager.AuthorizationCode))
-            //    {
-            //        App.ApiService = new MvpApiService(_loginManager.AuthorizationCode);
-
-            //        IsBusyMessage = "downloading profile info...";
-            //        Mvp = await App.ApiService.GetProfileAsync();
-
-            //        IsBusyMessage = "downloading profile image...";
-            //        ProfileImagePath = await App.ApiService.DownloadAndSaveProfileImage();
-
-            //        IsLoggedIn = true;
-            //    }
-            //}
-            //catch (Exception e)
-            //{
-            //    await e.LogExceptionAsync();
-            //}
-
-            //return !string.IsNullOrEmpty(_loginManager.AuthorizationCode);
-        }
-
-        public async Task SignOutAsync()
-        {
-            // Invoke sign out procedure (this will handle token storage internally), returns Tuple with success and message
-            //var result = await _loginManager.SignOutAsync();
-
-            try
-            {
-                IsLoggedIn = false;
-
-                var builder = new UriBuilder("https://login.live.com/oauth20_logout.srf");
-
-                var query = HttpUtility.ParseQueryString(builder.Query);
-                query["client_id"] = _clientId;
-                query["redirect_uri"] = _redirectUrl;
-                builder.Query = query.ToString();
-
-                WebViewUri = new Uri(builder.ToString());
-
-                StorageHelpers.Instance.DeleteToken("access_token");
-                StorageHelpers.Instance.DeleteToken("refresh_token");
-
-                // Clean up profile objects
-                IsLoggedIn = false;
-                Mvp = null;
-                ProfileImagePath = "";
-
-                // Clean up files
-                try
-                {
-                    // Profile photo file
-                    var imageFile = await ApplicationData.Current.LocalFolder.TryGetItemAsync("ProfilePicture.jpg");
-                    if (imageFile != null)
-                    {
-                        await imageFile.DeleteAsync(StorageDeleteOption.PermanentDelete);
-                    }
-
-                    // TODO Clean up Log files (there are automatically kept to a limit of 5 by the logging system).
-                    //var query = ApplicationData.Current.LocalFolder.CreateFileQuery();
-                    //var files = await query.GetFilesAsync();
-
-                    //foreach (var file in files)
-                    //{
-                    //    if (file.FileType == "log")
-                    //    {
-                    //        await file.DeleteAsync(StorageDeleteOption.PermanentDelete);
-                    //    }
-                    //}
-                }
-                catch (Exception e)
-                {
-                    await e.LogExceptionWithUserMessage();
-                    Debug.WriteLine(e);
-                }
-            }
-            catch (Exception ex)
-            {
-                await ex.LogExceptionWithUserMessage();
-            }
-
-            // If sign out was successful, handle app-specific files and resources
-            //if (result.Item1)
-            //{
-            //    // Clean up profile objects
-            //    IsLoggedIn = false;
-            //    Mvp = null;
-            //    ProfileImagePath = "";
-                
-            //    // Clean up files
-            //    try
-            //    {
-            //        // Profile photo file
-            //        var imageFile = await ApplicationData.Current.LocalFolder.TryGetItemAsync("ProfilePicture.jpg");
-            //        if (imageFile != null)
-            //        {
-            //            await imageFile.DeleteAsync(StorageDeleteOption.PermanentDelete);
-            //        }
-
-            //        // TODO Clean up Log files (there are automatically kept to a limit of 5 by the logging system).
-            //        //var query = ApplicationData.Current.LocalFolder.CreateFileQuery();
-            //        //var files = await query.GetFilesAsync();
-                    
-            //        //foreach (var file in files)
-            //        //{
-            //        //    if (file.FileType == "log")
-            //        //    {
-            //        //        await file.DeleteAsync(StorageDeleteOption.PermanentDelete);
-            //        //    }
-            //        //}
-            //    }
-            //    catch (Exception e)
-            //    {
-            //        await ExceptionLogger.LogExceptionAsync(e);
-            //        Debug.WriteLine(e);
-            //    }
-            //}
-
-            //// show user result of sign out attempt
-            //await new MessageDialog($"{result.Item2}", result.Item1 ? "Success" : "Error").ShowAsync();
-        }
-
-        public override async Task OnNavigatedToAsync(object parameter, NavigationMode mode, IDictionary<string, object> state)
-        {
-            if (!NetworkHelper.Instance.ConnectionInformation.IsInternetAvailable)
-            {
-                await new MessageDialog("This application requires an internet connection. Please check your connection and launch the app again.", "No Internet").ShowAsync();
-                return;
-            }
-        }
-
-        public override Task OnNavigatedFromAsync(IDictionary<string, object> pageState, bool suspending)
-        {
-            return base.OnNavigatedFromAsync(pageState, suspending);
-        }
-
-        #region Authentication
-
-        
-        //private readonly Uri _signOutUri = new Uri($"https://login.live.com/oauth20_logout.srf?client_id={_clientId}&redirect_uri=https:%2F%2Flogin.live.com%2Foauth20_desktop.srf");
+        #region UIElement event handlers
 
         public async void LoginMenuFlyoutItem_OnClick(object sender, RoutedEventArgs e)
         {
-            if (IsLoggedIn)
+            if (!IsLoggedIn)
             {
-                var md = new MessageDialog("You are already signed in. In case you're having trouble, you can force a refresh by logging out.");
-
+                await SignInAsync();
+            }
+            else
+            {
+                var md = new MessageDialog("You are already logged in. would you like to logout?");
                 md.Commands.Add(new UICommand("Logout"));
                 md.Commands.Add(new UICommand("Cancel"));
 
@@ -267,36 +97,11 @@ namespace MvpApi.Uwp.ViewModels
                     await SignOutAsync();
                 }
             }
-            else
-            {
-                // Is already logged out, try reloading the login url
-                WebViewUri = _signInUrl;
-            }
         }
 
         public async void LogoutMenuFlyoutItem_OnClick(object sender, RoutedEventArgs e)
         {
-            try
-            {
-                await SignOutAsync();
-                //IsLoggedIn = false;
-
-                //var builder = new UriBuilder("https://login.live.com/oauth20_logout.srf");
-
-                //var query = HttpUtility.ParseQueryString(builder.Query);
-                //query["client_id"] = _clientId;
-                //query["redirect_uri"] = _redirectUrl;
-                //builder.Query = query.ToString();
-
-                //WebViewUri = new Uri(builder.ToString());
-
-                //StorageHelpers.Instance.DeleteToken("access_token");
-                //StorageHelpers.Instance.DeleteToken("refresh_token");
-            }
-            catch (Exception ex)
-            {
-                await ex.LogExceptionWithUserMessage();
-            }
+            await SignOutAsync();
         }
 
         public async void WebView_OnNavigationCompleted(WebView sender, WebViewNavigationCompletedEventArgs e)
@@ -317,6 +122,84 @@ namespace MvpApi.Uwp.ViewModels
             {
                 // Redirect to signin page if there's a bounce
                 WebViewUri = _signInUrl;
+            }
+        }
+
+        #endregion
+        
+        #region Authentication
+
+        /// <summary>
+        /// Logs in the user by showing a ContentDialog for authentication.
+        /// Silent Operation if the user has previously logged in, a silent login will be attempted first (using the refresh token).
+        /// </summary>
+        /// <returns></returns>
+        public async Task SignInAsync()
+        {
+            var refreshToken = StorageHelpers.Instance.LoadToken("refresh_token");
+
+            if (!string.IsNullOrEmpty(refreshToken))
+            {
+                // there is a token stored, let's try to use it and not even have to show UI
+                await RequestAuthorizationAsync(refreshToken, true);
+            }
+            else
+            {
+                // no stored refresh_token was available, show dialog to get user to signin and accept
+
+                await TaskUtilities.RunOnDispatcherThreadAsync(() =>
+                {
+                    // Make sure the UI is showing login overlay
+                    if (IsLoggedIn)
+                        IsLoggedIn = false;
+
+                    WebViewUri = _signInUrl;
+                });
+            }
+        }
+
+        public async Task SignOutAsync()
+        {
+            try
+            {
+                // Erase cached tokens
+                StorageHelpers.Instance.DeleteToken("access_token");
+                StorageHelpers.Instance.DeleteToken("refresh_token");
+
+                // Clean up profile objects
+                Mvp = null;
+                ProfileImagePath = "";
+
+                try
+                {
+                    // Profile photo file
+                    var imageFile = await ApplicationData.Current.LocalFolder.TryGetItemAsync("ProfilePicture.jpg");
+                    if (imageFile != null)
+                    {
+                        await imageFile.DeleteAsync(StorageDeleteOption.PermanentDelete);
+                    }
+                }
+                catch (Exception e)
+                {
+                    await e.LogExceptionWithUserMessage();
+                    Debug.WriteLine($"File Reset Exception: {e}");
+                }
+            }
+            catch (Exception ex)
+            {
+                await ex.LogExceptionWithUserMessage();
+            }
+            finally
+            {
+                // Toggle the flag (this setter will trigger UI updates, including the showing the WebView to verify logout and navigate back to login page)
+                IsLoggedIn = false;
+
+                var builder = new UriBuilder("https://login.live.com/oauth20_logout.srf");
+                var query = HttpUtility.ParseQueryString(builder.Query);
+                query["client_id"] = _clientId;
+                query["redirect_uri"] = _redirectUrl;
+                builder.Query = query.ToString();
+                WebViewUri = new Uri(builder.ToString());
             }
         }
 
@@ -361,11 +244,13 @@ namespace MvpApi.Uwp.ViewModels
                         var tokenType = tokenData["token_type"];
                         var cleanedAccessToken = tokenData["access_token"].Split('&')[0];
 
-                        // set public property that is "returned"
+                        // Build the Bearer authorization header
                         var authorization = $"{tokenType} {cleanedAccessToken}";
 
+                        // New-up the service
                         App.ApiService = new MvpApiService(authorization);
 
+                        // Trigger UI changes (e.g. hide the overlay)
                         IsLoggedIn = true;
 
                         IsBusyMessage = "downloading profile info...";
@@ -376,7 +261,6 @@ namespace MvpApi.Uwp.ViewModels
 
                         // The current page should be HomePage, so we can just call refresh
                         BootStrapper.Current.NavigationService.Refresh();
-                        IsLoggedIn = true;
                     }
                 }
             }
@@ -390,16 +274,34 @@ namespace MvpApi.Uwp.ViewModels
                 }
 
                 Debug.WriteLine($"LoginDialog HttpRequestException: {e}");
-                
             }
             catch (Exception e)
             {
                 await e.LogExceptionWithUserMessage();
-
                 Debug.WriteLine($"LoginDialog Exception: {e}");
             }
         }
 
+        #endregion
+
+        #region Navigation
+    
+        public override async Task OnNavigatedToAsync(object parameter, NavigationMode mode, IDictionary<string, object> state)
+        {
+            if (!NetworkHelper.Instance.ConnectionInformation.IsInternetAvailable)
+            {
+                await new MessageDialog("This application requires an internet connection. Please check your connection and launch the app again.", "No Internet").ShowAsync();
+                return;
+            }
+
+            await SignInAsync();
+        }
+
+        public override Task OnNavigatedFromAsync(IDictionary<string, object> pageState, bool suspending)
+        {
+            return base.OnNavigatedFromAsync(pageState, suspending);
+        }
+        
         #endregion
     }
 }
